feat: Add RSA encryption and decryption methods
The `encrypt` method in `RSAEncryptor` now takes a `data` parameter instead of a `message` parameter and throws an exception if the data is too long to be encrypted. The `decrypt` method in `RSADecryptor` takes a `data` parameter instead of an `encryptedMessage` parameter and throws an exception if an error occurs during decryption.

The `encrypt` method in `RSAEncryptionHandler` now takes a `key` parameter instead of a `publicKey` parameter and uses the session's encrypt method if no key is provided. The `decrypt` method in `RSAEncryptionHandler` now takes a `key` parameter and uses the session's decrypt method.